Addition of common AbstractLocalTransformer class.
